+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - EAN13 validation rule
 - UPC-A validation rule
 - UPC-E validation rule
+- JAN (Japanese Article Number) validation rule
 
 ### Changed